requirements.txt
certifi==2019.3.9
chardet==3.0.4
gevent==1.4.0
greenlet==0.4.15
idna==2.8
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.55.0
